split product model
Splited product model into 2 models:
product
productMeta

also removed schema Specs and changed specs type to Mixed
