+++ server/models/product.js
@@ -2,11 +2,6 @@ const mongoose = require('mongoose');
 const Schema = mongoose.Schema;
 
 // Create Schema and Model
-const specSchema = new Schema({
-    title: String,
-    desc: String
-});
-
 const ProductSchema = new Schema({
     name: String,
     type: String,
@@ -15,20 +10,10 @@ const ProductSchema = new Schema({
     description: String,
     discount: Number,
     quantity: Number,
-    specs: [specSchema],
+    specs: [Schema.Types.Mixed],
     showInStore: Boolean
 });
 
-const ProductMetaSchema = new Schema({
-    name: String,
-    type: String,
-    company: String,
-    importerPrice: Number,
-    specs: [specsSchema]
-});
-
 const Product = mongoose.model('product', ProductSchema);
-const ProductMeta = mongoose.model('productMeta', ProductMetaSchema);
 
 module.exports.Product = Product;
-module.exports.ProductMeta = ProductMeta;
